In airplain

+++ java-project/src/main/java/com/yourself/Factorial.java
@@ -2,7 +2,7 @@ package com.yourself;
 
 public class Factorial {
 
-    public static int factorial (int n) {
+    public static long factorial (int n) {
         if (n == 0 || n==1) {
             return 1;
         } else {
